# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI 驅動的網紅案件管理系統 API\n提供 Google OAuth 認證、郵件管理、案件管理、AI 分析等功能",
        "title": "Influenter API",
        "termsOfService": "http://influenter.example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://influenter.example.com/support",
            "email": "support@influenter.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/google": {
            "post": {
                "description": "使用 Google credential 登入系統，如果使用者不存在會自動建立帳號",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "認證"
                ],
                "summary": "Google OAuth 登入",
                "parameters": [
                    {
                        "description": "Google 登入請求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.GoogleLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登入成功，返回使用者資訊和 JWT token",
                        "schema": {
                            "$ref": "#/definitions/github_com_designcomb_influenter-backend_internal_services.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "請求格式錯誤",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Google token 無效或驗證失敗",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器內部錯誤",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登出當前使用者，客戶端需清除本地儲存的 token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "認證"
                ],
                "summary": "使用者登出",
                "responses": {
                    "200": {
                        "description": "登出成功訊息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "未認證",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 JWT token 取得當前登入的使用者完整資訊",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "認證"
                ],
                "summary": "取得當前使用者資訊",
                "responses": {
                    "200": {
                        "description": "使用者資訊",
                        "schema": {
                            "$ref": "#/definitions/github_com_designcomb_influenter-backend_internal_models.User"
                        }
                    },
                    "401": {
                        "description": "未認證或 token 無效",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "使用者不存在",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器內部錯誤",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/emails": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取得使用者的郵件列表，支援分頁、篩選、搜尋",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "郵件"
                ],
                "summary": "取得郵件列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth 帳號 ID",
                        "name": "oauth_account_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否已讀",
                        "name": "is_read",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "案件 ID",
                        "name": "case_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "寄件者 email",
                        "name": "from_email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "主旨關鍵字",
                        "name": "subject",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "開始日期 (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "結束日期 (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "頁數",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每頁數量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "received_at",
                        "description": "排序欄位",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "description": "排序方向 (asc/desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "郵件列表和分頁資訊",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/emails/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取得單封郵件的完整內容",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "郵件"
                ],
                "summary": "取得郵件詳情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "郵件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_designcomb_influenter-backend_internal_models.EmailDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新郵件狀態（標記已讀、關聯案件等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "郵件"
                ],
                "summary": "更新郵件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "郵件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新內容",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.UpdateEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_designcomb_influenter-backend_internal_models.EmailDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gmail/disconnect": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "刪除 Gmail OAuth 帳號連接（保留已同步的郵件）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gmail"
                ],
                "summary": "斷開 Gmail 連接",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gmail/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取得使用者的 Gmail 帳號連接和同步狀態",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gmail"
                ],
                "summary": "取得 Gmail 同步狀態",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gmail/sync": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "手動觸發郵件同步（有冷卻時間限制）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gmail"
                ],
                "summary": "手動觸發 Gmail 同步",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "同步冷卻中",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_designcomb_influenter-backend_internal_models.EmailDetailResponse": {
            "type": "object",
            "properties": {
                "ai_analysis_id": {
                    "type": "string"
                },
                "ai_analyzed": {
                    "type": "boolean"
                },
                "body_html": {
                    "type": "string"
                },
                "body_text": {
                    "type": "string"
                },
                "case_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "from_email": {
                    "type": "string"
                },
                "from_name": {
                    "type": "string"
                },
                "has_attachments": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_read": {
                    "type": "boolean"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "oauth_account_id": {
                    "type": "string"
                },
                "provider_message_id": {
                    "type": "string"
                },
                "received_at": {
                    "type": "string"
                },
                "snippet": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "thread_id": {
                    "type": "string"
                },
                "to_email": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_designcomb_influenter-backend_internal_models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "profile_picture_url": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_designcomb_influenter-backend_internal_services.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_designcomb_influenter-backend_internal_models.User"
                }
            }
        },
        "internal_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_api.GoogleLoginRequest": {
            "type": "object",
            "required": [
                "credential"
            ],
            "properties": {
                "clientId": {
                    "type": "string"
                },
                "credential": {
                    "type": "string"
                }
            }
        },
        "internal_api.UpdateEmailRequest": {
            "type": "object",
            "properties": {
                "case_id": {
                    "type": "string"
                },
                "is_read": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}